+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog of lizard-nxt client
 Unreleased (2.1.0) (XXXX-XX-XX)
 ---------------------
 
+- Use transifex.com/api/2/lizard-client for translations. Push annotated
+and pull translated strings from transifex on grunt:build. Jenkins will keep
+transifex up to date while every release will use the newest strings.
+
 - Annotate waterchain attributes for translation and use translation filter in
 template.